requirements.txt
Flask==2.3.5
mysql-connector-python==8.1.0
